Add OAI 3.2 features, more bug fixes (#3)
* Fix codegen issues found via unhinged spec

- Escape $, control chars, and bidi codepoints in generated string
  literals (escapeDartString) so output always parses and formats
- Escape parameter keys in query/header map entries via escapeDartString
- Rename fields that shadow Object members (toString, hashCode, etc.)
  by extending sanitizeFieldName with objectMemberNames check
- Strip leading +/- before letters in enum value names to prevent
  casing corruption (+NaN → naN, not NaN)
- Integrate dart_style formatting into snapshot tests for unhinged and
  totem-mobile so formatting regressions are caught automatically

* Fix discriminator mapping, content-based params, and Latin transliteration

- Inferred oneOf discriminator mappings now use original schema names
  as switch keys instead of Dart type names
- Extract parameter schemas from OAS 3.x `content` field when `schema`
  is absent, fixing required-dynamic contradiction
- Transliterate accented Latin characters to ASCII (café→cafe) instead
  of stripping them (café→caf)

* Add integer/number enum support and improve canParse

- Extend IrEnum with valueKind to support int and double enums, not
  just string enums. Enum emitter generates typed value fields,
  fromJson/toJson with correct types, and deduplicates switch cases.
- canParse now checks for at least one known property key when a model
  has no required fields, instead of returning true unconditionally.

* Add integer/number enum support, improve canParse, document allOf limitation

- Extend IrEnum with valueKind to support int and double enums, not
  just string enums. Enum emitter generates typed value fields,
  fromJson/toJson with correct types, and deduplicates switch cases.
- canParse now checks for at least one known property key when a model
  has no required fields, instead of returning true unconditionally.
- Add test documenting allOf $ref+properties limitation (produces
  IrTypeRef instead of merged IrObject). Fixing this requires careful
  handling of discriminated union variants that use the same pattern.

* Add braces to if-case statements to satisfy curly_braces lint

* Extract shared emitter helpers for operator==, hashCode, toString, and primitiveKind codecs

- Add buildEqualsOverride, buildHashCodeOverride, buildToStringOverride
  to emit_utils.dart and use them across model, enum, and sealed union
  emitters (removes ~80 lines of identical scaffolding)
- Add primitiveFromJsonExpr and primitiveToJsonExpr to consolidate
  PrimitiveKind conversion logic in one place
- Reuse primitiveToJsonExpr in extension_type_emitter for toJson

* Add skipped tests documenting allOf ref+properties expected behavior

The allOf fix (resolving $ref targets during flattening) works correctly
in isolation but causes cascading issues across ~3200 Cloudflare allOf
schemas: self-referencing fromJson, undefined types from expanded aliases,
and field count mismatches. Tests document the expected behavior for
future implementation.

* Settings

* Expand allOf $ref + extra properties into merged IrObject

When allOf composes a $ref with additional inline properties, merge the
ref target's properties into the result instead of returning a bare
IrTypeRef that loses the extra fields. Discriminator variants (detected
via discriminatorProperty) are excluded from expansion to preserve
sealed union wrapping.

Also fixes: $-prefixed field names in sealed union variant toString,
and nullable toJson in object variants.

* Add --unwrap-fields CLI option for response envelope unwrapping

When --unwrap-fields=result is passed, API methods return the type of
the specified field directly instead of the full response envelope.
For Cloudflare's {success, errors, messages, result: T} pattern, this
means methods return T instead of the wrapper — matching what Stainless
generates for official Cloudflare SDKs.

The Cloudflare snapshot now uses unwrap_fields: [result], producing
cleaner API signatures like Future<ApiResult<Zone, Never>> instead of
Future<ApiResult<ZonesGetResponse, Never>>.

* Support OAS 3.2 query method + additionalOperations, fix discriminator toJson

- Add `query` as a first-class HTTP method (OAS 3.2.0)
- Parse `additionalOperations` map for custom HTTP methods (HAUNT, PURGE, etc.)
- Add HttpMethod.custom with customMethod string for arbitrary methods
- Fix discriminator toJson key collision: spread variant first, then set
  discriminator key so it always wins over variant's own type field
- Escape Object member names in operation method names (toString → $toString)

* Handle type: "null" as nullable dynamic instead of String

JSON Schema `type: "null"` means the only valid value is null.
Previously it fell through to the default case and was treated as
String, causing valid null payloads to fail parsing.

* Fix OAuth2 deviceAuthorization flow and type: null handling

- Add deviceAuthorization to ApiOAuthFlowType enum and flow mapping
- Add deviceAuthorizationUrl field to ApiOAuthFlow (runtime + IR)
- type: "null" now maps to nullable dynamic instead of String

* Add additionalProperties support for object models

Objects with both fixed properties and additionalProperties now generate
an overflow Map<String, T> field. fromJson separates known keys (fixed
fields) from unknown keys (overflow), and toJson spreads them back.

- Add additionalProperties field to IrObject IR type
- Populate it during schema lowering (true → dynamic, typed → T, false → null)
- ModelEmitter generates: overflow field, constructor with const {} default,
  fromJson filtering, toJson spreading, copyWith, mapEquals in ==, hashCode
- Add mapEquals to degenerate_runtime
- Resolve additionalProperties type refs in TypeRefResolver
- Update file_emitter import analysis for overflow map types

* Update example

* Add typed streaming support for itemSchema (SSE + JSONL)

- Parse itemSchema from OpenAPI media type objects during response lowering
- SSE streams with itemSchema return Stream<T> with typed deserialization
- Add JSONL (application/jsonl, application/x-ndjson) streaming support
  with parseJsonlStream in runtime and executeJsonlStreaming in executor
- Prefer itemSchema over schema for streaming event type
- Add StreamKind enum to distinguish SSE from JSONL in emitter

Fixes: streamBarnEvents now returns Stream<TypeModel> and $yield returns
Stream<StringModel>, instead of Future<ApiResult<void, Never>>.

* Update README and CHANGELOG for v0.3.0

Document all new features: additionalProperties, allOf composition,
--unwrap-fields, OAS 3.2 support, typed streaming, integer enums,
and bug fixes. Add Unhinged to tested specs table.

+++ snapshots/public/cloudflare/lib/models/certificate_packs_order_advanced_certificate_manager_certificate_pack_request.dart
@@ -30,7 +30,7 @@ factory CertificatePacksOrderAdvancedCertificateManagerCertificatePackRequest.fr
   hosts: (json['hosts'] as List<dynamic>).map((e) => e as String).toList(),
   type: TlsCertificatesAndHostnamesAdvancedType.fromJson(json['type'] as String),
   validationMethod: TlsCertificatesAndHostnamesValidationMethod.fromJson(json['validation_method'] as String),
-  validityDays: TlsCertificatesAndHostnamesValidityDays.fromJson(json['validity_days'] as num),
+  validityDays: TlsCertificatesAndHostnamesValidityDays.fromJson((json['validity_days'] as num).toInt()),
 ); }
 
 /// Certificate Authority selected for the order.  For information on any certificate authority specific details or restrictions [see this page for more details.](https://developers.cloudflare.com/ssl/reference/certificate-authorities)
