+++ snapshots/public/cloudflare/lib/models/origin_ca_create_certificate_request.dart
@@ -6,7 +6,7 @@ factory OriginCaCreateCertificateRequest.fromJson(Map<String, dynamic> json) { r
   csr: TlsCertificatesAndHostnamesCsr.fromJson(json['csr'] as String),
   hostnames: (json['hostnames'] as List<dynamic>).map((e) => e as String).toList(),
   requestType: TlsCertificatesAndHostnamesRequestType.fromJson(json['request_type'] as String),
-  requestedValidity: json['requested_validity'] != null ? TlsCertificatesAndHostnamesRequestedValidity.fromJson(json['requested_validity'] as num) : null,
+  requestedValidity: json['requested_validity'] != null ? TlsCertificatesAndHostnamesRequestedValidity.fromJson((json['requested_validity'] as num).toDouble()) : null,
 ); }
 
 /// The Certificate Signing Request (CSR). Must be newline-encoded.
